+++ lib/mapospace_flutter_sdk.dart
@@ -1,4 +1,6 @@
 /// **Mapospace Flutter SDK** is a lightweight geospatial analytics tool that captures event data, including location, from mobile apps for real-time insights. It provides easy integration and powerful location-based analytics to help understand user behavior and trends.
 library;
 
+export 'src/events/sale_event.dart';
 export 'src/mapospace_flutter_sdk.dart';
+export 'src/models/product_model.dart';